+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/Light/HDAdditionalLightData.cs
@@ -117,7 +117,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
         [Range(0.0f, 1.0f), SerializeField, FormerlySerializedAs("volumetricDimmer")]
         private float m_VolumetricDimmer = 1.0f;
-        
+
         public float volumetricDimmer
         {
             get { return useVolumetric ? m_VolumetricDimmer : 0f; }
@@ -143,7 +143,9 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public LightTypeExtent lightTypeExtent = LightTypeExtent.Punctual;
 
         // Only for Spotlight, should be hide for other light
-        public SpotLightShape spotLightShape = SpotLightShape.Cone;
+        public SpotLightShape spotLightShape { get { return m_SpotLightShape; } set { SetSpotLightShape(value); } }
+        [SerializeField, FormerlySerializedAs("spotLightShape")]
+        SpotLightShape m_SpotLightShape = SpotLightShape.Cone;
 
         // Only for Rectangle/Line/box projector lights
         public float shapeWidth = 0.5f;
@@ -177,7 +179,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
         // Optional cookie for rectangular area lights
         public Texture areaLightCookie = null;
-        
+
         // Duplication of HDLightEditor.k_MinAreaWidth, maybe do something about that
         const float k_MinAreaWidth = 0.01f; // Provide a small size of 1cm for line light
 
@@ -253,7 +255,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         {
             Bounds bounds;
             float cameraDistance = Vector3.Distance(camera.transform.position, transform.position);
-            
+
             #if ENABLE_RAYTRACING
             m_WillRenderShadows = m_Light.shadows != LightShadows.None && frameSettings.IsEnabled(FrameSettingsField.Shadow) && lightTypeExtent == LightTypeExtent.Punctual;
             #else
@@ -269,7 +271,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
             if (!m_WillRenderShadows)
                 return;
-            
+
             // Create shadow requests array using the light type
             if (shadowRequests == null || m_ShadowRequestIndices == null)
             {
@@ -280,7 +282,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 for (int i = 0; i < maxLightShadowRequestsCount; i++)
                     shadowRequests[i] = new HDShadowRequest();
             }
-            
+
             Vector2 viewportSize = new Vector2(m_ShadowData.shadowResolution, m_ShadowData.shadowResolution);
 
             // Compute dynamic shadow resolution
@@ -748,6 +750,76 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 #endif
         }
 
+        void SetSpotLightShape(SpotLightShape shape)
+        {
+            m_SpotLightShape = shape;
+            UpdateBounds();
+        }
+
+        void UpdateAreaLightBounds()
+        {
+            m_Light.useShadowMatrixOverride = false;
+            m_Light.useBoundingSphereOverride = true;
+            m_Light.boundingSphereOverride = new Vector4(0.0f, 0.0f, 0.0f, m_Light.range);
+        }
+
+        void UpdateBoxLightBounds()
+        {
+            m_Light.useShadowMatrixOverride = true;
+            m_Light.useBoundingSphereOverride = true;
+
+            // Need to inverse scale because culling != rendering convention apparently
+            Matrix4x4 scaleMatrix = Matrix4x4.Scale(new Vector3(1.0f, 1.0f, -1.0f));
+            m_Light.shadowMatrixOverride = HDShadowUtils.ExtractBoxLightProjectionMatrix(m_Light.range, shapeWidth, shapeHeight, shadowNearPlane) * scaleMatrix;
+
+            // Very conservative bounding sphere taking the diagonal of the shape as the radius
+            float diag = new Vector3(shapeWidth * 0.5f, shapeHeight * 0.5f, m_Light.range * 0.5f).magnitude;
+            m_Light.boundingSphereOverride = new Vector4(0.0f, 0.0f, m_Light.range * 0.5f, diag);
+        }
+
+        void UpdatePyramidLightBounds()
+        {
+            m_Light.useShadowMatrixOverride = true;
+            m_Light.useBoundingSphereOverride = true;
+
+            // Need to inverse scale because culling != rendering convention apparently
+            Matrix4x4 scaleMatrix = Matrix4x4.Scale(new Vector3(1.0f, 1.0f, -1.0f));
+            m_Light.shadowMatrixOverride = HDShadowUtils.ExtractSpotLightProjectionMatrix(m_Light.range, m_Light.spotAngle, shadowNearPlane, aspectRatio, 0.0f) * scaleMatrix;
+
+            // Very conservative bounding sphere taking the diagonal of the shape as the radius
+            float diag = new Vector3(shapeWidth * 0.5f, shapeHeight * 0.5f, m_Light.range * 0.5f).magnitude;
+            m_Light.boundingSphereOverride = new Vector4(0.0f, 0.0f, m_Light.range * 0.5f, diag);
+        }
+
+        void UpdateBounds()
+        {
+            if (lightTypeExtent == LightTypeExtent.Punctual && m_Light.type == LightType.Spot)
+            {
+                switch (spotLightShape)
+                {
+                    case SpotLightShape.Box:
+                        UpdateBoxLightBounds();
+                        break;
+                    case SpotLightShape.Pyramid:
+                        UpdatePyramidLightBounds();
+                        break;
+                    default: // Cone
+                        m_Light.useBoundingSphereOverride = false;
+                        m_Light.useShadowMatrixOverride = false;
+                        break;
+                }
+            }
+            else if (lightTypeExtent == LightTypeExtent.Rectangle || lightTypeExtent == LightTypeExtent.Tube)
+            {
+                UpdateAreaLightBounds();
+            }
+            else
+            {
+                m_Light.useBoundingSphereOverride = false;
+                m_Light.useShadowMatrixOverride = false;
+            }
+        }
+
         // As we have our own default value, we need to initialize the light intensity correctly
         public static void InitDefaultHDAdditionalLightData(HDAdditionalLightData lightData)
         {
@@ -788,7 +860,15 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             light.lightShadowCasterMode = LightShadowCasterMode.Everything;
         }
 
-        public void OnBeforeSerialize() { }
+        void OnValidate()
+        {
+            UpdateBounds();
+        }
+
+        public void OnBeforeSerialize()
+        {
+            UpdateBounds();
+        }
 
         public void OnAfterDeserialize()
         {